Code clean-up and general organisation

+++ Booking System/frontend/user/bookings/modify/ListBookings.cs
@@ -12,13 +12,16 @@ namespace Booking_System.frontend.user.bookings.modify
     {
         private Booking[] bookings;
         private User user;
+        private bool fromAdmin;
 
-        public ListBookings(User user)
+        public ListBookings(User user, bool fromAdmin=false)
         {
             InitializeComponent();
             this.user = user;
+            this.fromAdmin = fromAdmin;
             LoadBookingList();
         }
+
 
         private void LoadBookingList()
         {
@@ -35,12 +38,19 @@ namespace Booking_System.frontend.user.bookings.modify
                     Hotel hotel = HotelWrapper.GetHotel(room.HotelId);
                     Debug.WriteLine(hotel.Name);
 
+                    //Set the list of hotels
                     ListViewItem listViewItem = new ListViewItem(bookings[i].Id + "");
                     listViewItem.SubItems.Add(hotel.Name);
                     listViewItem.SubItems.Add(room.Name);
                     listViewItem.SubItems.Add(bookings[i].GetAmountOfNights() + "");
                     listViewItem.SubItems.Add($"£{bookings[i].Price}");
+
+                    char editable = 'N';
+                    if (bookings[i].IsBeforeCheckIn()) editable = 'Y';
+
+                    listViewItem.SubItems.Add($"{editable}");
                     listViewBookings.Items.Add(listViewItem);
+                    //Set the list of hotels
                 }
             }
             catch (Exception ex)
@@ -63,17 +73,42 @@ namespace Booking_System.frontend.user.bookings.modify
 
         private void buttonViewBooking_Click(object sender, EventArgs e)
         {
+            Booking booking = this.GetSelectedBooking();
+            if (booking == null) return;
 
+            if (!booking.IsBeforeCheckIn() && !this.fromAdmin)
+            {
+                this.ShowError("Unable to delete a booking which is AFTER check-in date. Please contact an administrator.");
+                return;
+            }
+
+            //Makes more sense like this since everything is in the cache already
+            Room room = RoomWrapper.GetHotelRoom(booking.RoomId);
+            Hotel hotel = HotelWrapper.GetHotel(room.HotelId);
+            //Makes more sense like this since everything is in the cache already
+
+            ModifyBooking viewBooking = new ModifyBooking(hotel, room, booking);
+            viewBooking.Show();
+            viewBooking.Closed += this.WindowBookingList_Close; //When the room is closed event is triggered, reload the booking list
         }
+
 
         private void buttonDeleteBooking_Click(object sender, EventArgs e)
         {
+            Booking booking = this.GetSelectedBooking();
+            if (booking == null) return;
+
+            if (!booking.IsBeforeCheckIn() && !this.fromAdmin)
+            {
+                this.ShowError("Unable to delete a booking which is AFTER check-in date.  Please contact an administrator.");
+                return;
+            }
+
             DialogResult dialogResult = MessageBox.Show($"Are you sure you want to delete this booking?",
                 "Confirm booking delete", MessageBoxButtons.YesNo);
             if (dialogResult != DialogResult.Yes) return; //If no, stop action
 
-            Booking booking = this.GetSelectedBooking();
-            if (booking == null) return;
+            
 
             BookingWrapper.DeleteBooking(booking);
             this.LoadBookingList();
@@ -83,5 +118,11 @@ namespace Booking_System.frontend.user.bookings.modify
         {
             MessageBox.Show(message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
         }
+
+        private void WindowBookingList_Close(object sender, EventArgs e)
+        {
+            this.LoadBookingList();
+        }
+
     }
 }
